refactor(api): rename training route file
