+++ api/src/devcomp/application/modules/module-controller.js
