+++ api/src/devcomp/application/trainings/training-route.js
